+++ balance.h
@@ -4,8 +4,9 @@
  * http://creativecommons.org/licenses/by-nc-nd/4.0/.
  */
 
+
 #include "gb.h"
 
-BOOLEAN has_money(unsigned long amount_h, unsigned long amount_l);
-BOOLEAN remove_money(unsigned long amount_h, unsigned long amount_l);
-void add_money(unsigned long amount_h, unsigned long amount_l);
+BOOLEAN has_money(UINT16 amount_h, UINT16 amount_l);
+BOOLEAN remove_money(UINT16 amount_h, UINT16 amount_l);
+void add_money(UINT16 amount_h, UINT16 amount_l);
